prepare controller abstraction
