+++ crates/controller/src/button/mod.rs
@@ -1,9 +1,8 @@
-mod control;
 mod error;
-mod event;
+mod state;
 
-use crate::keys::KEYBOARD_BUTTON_MAP;
-use evdev::KeyCode;
+pub use error::ButtonError;
+pub use state::State;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum ControllerButton {
@@ -29,45 +28,6 @@ pub enum ControllerButton {
   Select,
 }
 
-impl TryFrom<ControllerButton> for KeyCode {
-  type Error = ButtonError;
-
-  fn try_from(value: ControllerButton) -> Result<Self, Self::Error> {
-    let code = match value {
-      ControllerButton::South => KeyCode::BTN_SOUTH,
-      ControllerButton::East => KeyCode::BTN_EAST,
-      ControllerButton::North => KeyCode::BTN_WEST,
-      ControllerButton::West => KeyCode::BTN_NORTH,
-      ControllerButton::Up => KeyCode::BTN_DPAD_UP,
-      ControllerButton::Down => KeyCode::BTN_DPAD_DOWN,
-      ControllerButton::Left => KeyCode::BTN_DPAD_LEFT,
-      ControllerButton::Right => KeyCode::BTN_DPAD_RIGHT,
-      ControllerButton::L1 => KeyCode::BTN_TL,
-      ControllerButton::R1 => KeyCode::BTN_TR,
-      ControllerButton::L2 => KeyCode::BTN_TL2,
-      ControllerButton::R2 => KeyCode::BTN_TR2,
-      ControllerButton::L3 => KeyCode::BTN_THUMBL,
-      ControllerButton::R3 => KeyCode::BTN_THUMBR,
-      ControllerButton::Start => KeyCode::BTN_START,
-      ControllerButton::Select => KeyCode::BTN_SELECT,
-      _ => return Err(ButtonError::InvalidButton(value)),
-    };
-    Ok(code)
-  }
-}
-
-impl TryFrom<KeyCode> for ControllerButton {
-  type Error = ButtonError;
-
-  fn try_from(value: KeyCode) -> Result<Self, Self::Error> {
-    if let Some(button) = KEYBOARD_BUTTON_MAP.get(&value) {
-      Ok(*button)
-    } else {
-      Err(ButtonError::UnsupportedKeyCode(value))
-    }
-  }
-}
-
 #[rustfmt::skip]
 impl ControllerButton {
   pub fn all() -> &'static [Self] {
@@ -83,6 +43,3 @@ impl ControllerButton {
     ]
   }
 }
-
-pub use error::ButtonError;
-pub use event::ControllerButtonEvent;
