+++ crates/gui/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "controller"
+name = "gui"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
 
 [dependencies]
-settings = { path = "../settings" }
-io = { path = "../io" }
