+++ crates/platform/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "controller"
+name = "platform"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -8,3 +8,7 @@ authors.workspace = true
 [dependencies]
 settings = { path = "../settings" }
 io = { path = "../io" }
+
+[target.'cfg(not(windows))'.dependencies]
+epoll = "4.3.3"
+evdev = "0.13.1"
