+++ crates/platform/src/joystick/direction.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum Direction {
+  North,
+  NorthEast,
+  East,
+  SouthEast,
+  South,
+  SouthWest,
+  West,
+  NorthWest,
+}
