+++ crates/platform/src/joystick/error.rs
