+++ crates/platform/src/joystick/mod.rs
@@ -1,0 +1,66 @@
+mod axis;
+mod control;
+mod direction;
+mod error;
+mod event;
+mod keys;
+mod motion;
+mod polarity;
+mod state;
+mod vector;
+
+use evdev::{KeyCode, KeyEvent, RelativeAxisCode, RelativeAxisEvent};
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum JoyStick {
+  Left,
+  Right,
+}
+
+impl TryFrom<RelativeAxisCode> for JoyStick {
+  type Error = JoyStickError;
+
+  fn try_from(value: RelativeAxisCode) -> Result<Self, Self::Error> {
+    let joystick = match value {
+      RelativeAxisCode::REL_X => Self::Right,
+      RelativeAxisCode::REL_Y => Self::Right,
+      other => return Err(JoyStickError::UnsupportedAxisCode(other)),
+    };
+    Ok(joystick)
+  }
+}
+
+impl TryFrom<RelativeAxisEvent> for JoyStick {
+  type Error = JoyStickError;
+
+  fn try_from(value: RelativeAxisEvent) -> Result<Self, Self::Error> {
+    Self::try_from(value.code())
+  }
+}
+
+impl TryFrom<KeyCode> for JoyStick {
+  type Error = JoyStickError;
+
+  fn try_from(value: KeyCode) -> Result<Self, Self::Error> {
+    if !JOYSTICK_KEYS.code_is_joystick_key(value) {
+      return Err(JoyStickError::UnsupportedKeyCode(value));
+    }
+    Ok(JoyStick::Left)
+  }
+}
+
+impl TryFrom<KeyEvent> for JoyStick {
+  type Error = JoyStickError;
+
+  fn try_from(value: KeyEvent) -> Result<Self, Self::Error> {
+    Self::try_from(value.code())
+  }
+}
+
+pub use axis::{AxisError, JoyStickAxis};
+pub use direction::Direction;
+pub use error::JoyStickError;
+pub use event::ControllerJoyStickEvent;
+pub use keys::JOYSTICK_KEYS;
+pub use state::JoyStickState;
+pub use vector::Vector;
