+++ crates/platform/src/state.rs
